Сделал основное ТЗ + подгнал под критерии
