+++ project/src/components/app/app.tsx
@@ -1,13 +1,13 @@
 import {Route, Routes} from 'react-router-dom';
-import MainScreen from '../../pages/mainScreen/mainScreen';
-import LoginScreen from '../../pages/loginScreen/loginScreen';
-import FavotitesScreen from '../../pages/favoritesScreen/favoritesScreen';
-import NotFoundScreen from '../../pages/notFoundScreen/notFoundScreen';
-import RoomScreen from '../../pages/roomScreen/roomScreen';
+import MainScreen from '../../pages/main-screen/main-screen';
+import LoginScreen from '../../pages/login-screen/login-screen';
+import FavotitesScreen from '../../pages/favorites-screen/favorites-screen';
+import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
+import RoomScreen from '../../pages/room-screen/room-screen';
 import {AppRoute} from '../../const';
 import PrivateRoute from '../../components/private-route/private-route';
-import HistoryRouter from '../historyRoute/historyRoute';
-import { browserHistory } from '../../browserHistory';
+import HistoryRouter from '../history-route/history-route';
+import { browserHistory } from '../../browser-history';
 
 function App(): JSX.Element {
