+++ project/src/components/error-message/error-message.tsx
@@ -1,0 +1,12 @@
+import { useAppSelector } from '../../hooks';
+import { getErrorMessage } from '../../store/app-process/selectors';
+import './error-message.css';
+
+function ErrorMessage(): JSX.Element | null {
+  const error = useAppSelector(getErrorMessage);
+  return error
+    ? <div className="error-message">{error}</div>
+    : null;
+}
+
+export default ErrorMessage;
