--- project/src/components/errorMessage/errorMessage.tsx
@@ -1,12 +1,0 @@
-import { useAppSelector } from '../../hooks';
-import { getErrorMessage } from '../../store/appProcess/selectors';
-import './errorMessage.css';
-
-function ErrorMessage(): JSX.Element | null {
-  const error = useAppSelector(getErrorMessage);
-  return error
-    ? <div className="error-message">{error}</div>
-    : null;
-}
-
-export default ErrorMessage;
