+++ project/src/components/header-nav/header-nav.tsx
@@ -1,0 +1,14 @@
+import { AuthorizationStatus } from '../../const';
+import { useAppSelector } from '../../hooks';
+import { getAuthorizationStatus } from '../../store/user-process/selectors';
+import HeaderNavLogged from '../header-nav-logged/header-nav-logged';
+import HeaderNavNotLogged from '../header-nav-not-logged/header-nav-not-logged';
+
+function HeaderNav(): JSX.Element {
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
+  return authorizationStatus === AuthorizationStatus.Auth
+    ? <HeaderNavLogged />
+    : <HeaderNavNotLogged />;
+}
+
+export default HeaderNav;
