+++ project/src/pages/favorites-screen-empty/favorites-screen-empty.tsx
