+++ project/src/store/middlewares/redirect.ts
@@ -1,6 +1,6 @@
 import { Middleware } from '@reduxjs/toolkit';
-import { browserHistory } from '../../browserHistory';
-import { rootReducer } from '../rootReducer';
+import { browserHistory } from '../../browser-history';
+import { rootReducer } from '../root-reducer';
 
 type Reducer = ReturnType<typeof rootReducer>;
