+++ project/src/store/root-reducer.ts
@@ -1,0 +1,11 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import { NameSpace } from '../const';
+import { appProcess } from './app-process/app-process';
+import { dataProcess } from './data-process/data-process';
+import { userProcess } from './user-process/user-process';
+
+export const rootReducer = combineReducers({
+  [NameSpace.User]: userProcess.reducer,
+  [NameSpace.Data]: dataProcess.reducer,
+  [NameSpace.App]: appProcess.reducer,
+});
