+++ project/src/types/state.ts
@@ -13,11 +13,11 @@ export type UserProcess = {
 export type DataProcess = {
   offers: Offers;
   isLoaded: boolean;
-  nearbyOffers: Offers;
+  offersNearby: Offers;
   comments: Comments;
   offer: Offer | null;
-  postLoaded: boolean;
-  favoriteOffers: Offers;
+  isPostLoaded: boolean;
+  offersFavorites: Offers;
   reloadFavorites: boolean;
   commentPostStatus: boolean;
 };
